some re-org and testing
:heavy_check_mark: move subset to main.py for maf command
:heavy_check_mark: a few small argument and message changes
:heavy_check_mark: adding testing
:heavy_check_mark: re-orging inputs for testing
