+++ postprocessing_variant_calls/maf/main.py
@@ -2,6 +2,7 @@ from .annotate import annotate_process
 from .concat import concat_process
 from .concat.concat_helpers import concat_mafs, check_maf, check_txt, process_paths, process_header
 from .concat.resources import de_duplication_columns, minimal_maf_columns
+from .subset.subset_helpers import read_tsv, read_ids, filter_by_rows
 import typer 
 import logging
 from pathlib import Path
@@ -88,7 +89,70 @@ def maf_maf(
     # write out paths
     concat_df.to_csv(f"{output_maf}.maf", index=False, sep="\t")
     return 0
-    
+
+# Add Subset App 
+@app.command("subset", help="subset maf files")
+def subset_maf(
+    maf: Path = typer.Option(
+        None, 
+        "--maf",
+        "-m",
+        exists=True,
+        file_okay=True,
+        dir_okay=False,
+        writable=False,
+        readable=True,
+        resolve_path=True,
+        help="MAF file to subset",
+    ),
+    ids: Path = typer.Option(
+        "",
+        "--ids",
+        "-i",
+        help="List of ids to search for in the 'Tumor_Sample_Barcode' column. Header of this file is 'sample_id'",
+    ),
+    sid: Optional[List[str]] = typer.Option(
+        None,
+        "--sid",
+        help="Identifiers to search for in the 'Tumor_Sample_Barcode' column. Can be given multiple times",
+    ),
+    output_file: str = typer.Option(
+        "output_subset.maf",
+        "--output",
+        "-o",
+        help="Name of the output file",
+    ),
+    col_name: str = typer.Option(
+        "Tumor_Sample_Barcode",
+        "--cname",
+        "-c",
+        help="Name of the column header to be used for sub-setting",
+    ),
+):
+
+    """
+    Subset MAF/TSV file by selected sample ID/IDs
+
+    Tool to do the following operations:
+    A. Get subset of variants based on Tumor_Sample_Barcode in data_mutations_extended.txt file
+    B. Mark the variants as overlapping with BED file as covered [yes/no], by appending "covered" column to the subset MAF
+
+    Requirement:
+    pandas; typing; typer; bed_lookup(https://github.com/msk-access/python_bed_lookup)
+
+    """
+    typer.echo("Subset and Annotate MAF...")
+    if not ids:
+        typer.echo("Identifiers were not provided in a text file")
+        if not sid:
+            typer.echo("Identifiers were not provided via command line as well")
+            raise typer.Abort()
+
+    maf_df = read_tsv(maf)
+    ids_to_subset = read_ids(sid, ids)
+    subset_maf = filter_by_rows(ids_to_subset, maf_df, col_name)
+    subset_maf.drop_duplicates().to_csv(output_file, sep="\t", index=False)
+    typer.echo("Done!")
 
 if __name__ == "__main__":
     app()
