+++ resources/maf_subset/example_subset_ids.txt
@@ -1,0 +1,2 @@
+sample_id
+C-35XM6W-L002-d02
